+++ application/main/project/kbd.mk
@@ -27,6 +27,7 @@ SRC_FILES += $(APP_SRC_DIR)/keyboard/keyboard_matrix.c \
     $(APP_SRC_DIR)/protocol/hid_configuration.c \
     $(APP_SRC_DIR)/keyboard/sleep_reason.c \
     $(APP_SRC_DIR)/keyboard/keyboard_evt.c \
+    $(APP_SRC_DIR)/keyboard/keyboard_battery.c \
     $(APP_SRC_DIR)/keyboard/store_config.c \
     $(APP_SRC_DIR)/protocol/usb_comm.c \
     $(APP_SRC_DIR)/protocol/ble_comm.c \
